Moved areadetector tests back into the andor detector test class, they will stay there until they need to be moved
